restructure folders
Each command has its own folder now using an index file. This allows for each command to house subcommands and other utility files specific to them.
