+++ src/utilities/version.ts
@@ -1,2 +1,2 @@
 // generated by genversion
-module.exports = '2.1.0';
+export const version = '2.2.2';
